Move Api out of client.py into its own file
